requirements-light.txt
# REQUIREMENTS OPTIMIZADOS - SIN PYTORCH
# Sistema inteligente basado solo en patrones

# --- Backend web ---
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
Werkzeug==2.3.7

# --- Processing básico ---
numpy==1.24.4
Pillow==9.5.0
pandas==1.5.3

# --- OCR básico (solo Tesseract) ---
pytesseract==0.3.10

# --- PDF processing ---
pdf2image==1.16.3

# --- Utilities ---
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1

# --- Image processing básico ---
opencv-python-headless==4.8.1.78

# --- Text processing ---
python-bidi==0.4.0
six==1.16.0